requirements.txt
### dependency
tqdm
dvc
pandas
numpy
PyYAML
tensorflow-gpu
matplotlib
joblib


### local packages -
-e . # << uncomment this to install from setup.py
